app_gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
PyJWT
pydantic
pydantic-settings
